7257 - MultipageNavMobile component should not be react component
