+++ network-api/networkapi/wagtailpages/templates/buyersguide/pni_mobile_nav.html
@@ -1,6 +1,5 @@
 {% load bg_nav_tags localization i18n static wagtailroutablepage_tags %}
 
-<!-- TODO: Replace this with the multipage-nav-mobile once we change that from a react component to a vanilla one -->
 <div class="d-md-none mt-0 mb-0" id="pni-nav-mobile">
     <div class="container">
         <div class="row px-3 px-sm-0">
