+++ source/js/foundation/inject-react/index.js
@@ -1,9 +1,8 @@
-import injectMultipageNav from "./multipage-nav.js";
 import injectNews from "./news.js";
 import injectPetition from "./petition.js";
 import injectPulseProjectList from "./pulse-project-list.js";
 import injectShareButtonGroup from "./share-button-group.js";
-import injectPulseTabbedProfileDirectory from "./pulse-tabbed-profile-filters"
+import injectPulseTabbedProfileDirectory from "./pulse-tabbed-profile-filters";
 
 /**
  * Inject React components
@@ -12,7 +11,6 @@ import injectPulseTabbedProfileDirectory from "./pulse-tabbed-profile-filters"
  * @param {Object} env Object of environment variables
  */
 export const injectReactComponents = (apps, siteUrl, env) => {
-  injectMultipageNav(apps);
   injectNews(apps, env);
   injectPetition(apps, siteUrl);
   injectPulseProjectList(apps, env);
